+++ packages/blazor-workspace/NimbleBlazor/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/nimble-blazor
 
-<!-- This log was last generated on Wed, 18 Mar 2026 21:46:50 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 24 Mar 2026 22:49:20 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 20.14.3
+
+Tue, 24 Mar 2026 22:49:20 GMT
+
+### Patches
+
+- Minor README change to test nuget publishing ([ni/nimble@cc77ce0](https://github.com/ni/nimble/commit/cc77ce00918ffa501722ab524b60fd846ae20a76))
+
+## 20.14.2
+
+Thu, 19 Mar 2026 14:12:53 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v35.5.2
 
 ## 20.14.1