+++ packages/spright-components/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @ni/spright-components
 
-<!-- This log was last generated on Wed, 18 Mar 2026 21:46:50 GMT and should not be manually modified. -->
+<!-- This log was last generated on Mon, 23 Mar 2026 22:18:57 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 6.14.4
+
+Mon, 23 Mar 2026 22:18:57 GMT
+
+### Patches
+
+- replace field-sizing with js approach ([ni/nimble@2309573](https://github.com/ni/nimble/commit/230957398badde38a5f2fc5116cd7eef9a9cfdfa))
+
+## 6.14.3
+
+Thu, 19 Mar 2026 14:12:53 GMT
+
+### Patches
+
+- Bump fast-foundation version to 10.2.2 ([ni/nimble@452fa5a](https://github.com/ni/nimble/commit/452fa5a0cd2dfd16749eeeeecf4b3eea771719ae))
+- Bump @ni/nimble-components to v35.5.2
 
 ## 6.14.2